backend/requirements.txt
# Flask and web framework
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Telegram bot
python-telegram-bot==20.7

# AI Services
google-generativeai>=0.8.0
replicate==0.22.0
openai==1.6.1
requests==2.31.0

# Google Sheets
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0

# Audio/Video processing
pydub==0.25.1
Pillow==10.1.0

# Utilities
python-magic==0.4.27
validators==0.22.0
